backend/pyproject.toml
[project]
name = "locallingua-backend"
version = "0.1.0"
description = "FastAPI backend for LocalLingua"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.6",
  "langdetect>=1.0.9",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
llama = ["llama-cpp-python>=0.2.90"]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = ["--import-mode=prepend"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
